requirements.txt
anthropic==0.42.0
streamlit==1.40.2
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
openpyxl==3.1.5
python-dotenv==1.0.1
